[refactor/design] Divide component & apply component in today sentence page - apply like icon component - modify css to apply like icon - divide each sentence component
