+++ polaris/src/component/LikeIcon.js
@@ -30,9 +30,6 @@ const LikeIcon = () => {
 
 
 const Icon = styled.div`
-    margin-left: auto;
-    margin-right: 25px;
-    
 `;
 
 const HeartFilledIcon = styled(AiFillHeart)`
