+++ polaris/src/page/TodaySentencePage.js
@@ -2,21 +2,12 @@ import React, { useState } from 'react';
 import styled from "styled-components";
 import NavBar from "../component/NavBar";
 import FooterBar from "../component/FooterBar";
-import { PiBookFill } from "react-icons/pi";
-import { AiFillHeart, AiOutlineHeart } from "react-icons/ai";
 import Modal from 'react-modal';
+import LikeIcon from "../component/LikeIcon"; 
+import EachSentence from "../component/EachSentence"; 
 
 const TodaySentencePage = () => {
-    const [isChecked, setIsChecked] = useState(false);
     const [modalIsOpen, setModalIsOpen] = useState(false);
-    // like button
-    const onClick = () => {
-        if (isChecked) {
-            setIsChecked(false);
-        } else {
-            setIsChecked(true);
-        }
-    };
     return (
         <>
             <NavBar/>
@@ -25,30 +16,10 @@ const TodaySentencePage = () => {
                 <SubtitleText>오늘의 한 문장이 여러분 책여행의 북극성이 되어줄 것입니다.</SubtitleText>
                 <ExplainationText>※ 하루에 한 문장만 선택해서 책을 추천받을 수 있습니다.</ExplainationText>
                 <SentencesContainer className="container-fluid row">
-                    <EachSentence onClick={()=> setModalIsOpen(true)}>
-                        <QuoteBox >
-                            <QuoteText>고독을 배설한 자리에서 내려앉는 환희. 이 달콤함을 위해 그는 예술을 표방한다.</QuoteText>
-                        </QuoteBox >
-                        <BookIcon size="250" color="#97A4E8" />
-                    </EachSentence>
-                    <EachSentence onClick={()=> setModalIsOpen(true)}>
-                        <QuoteBox>
-                            <QuoteText>전부 바다에 밀어버리자. 더 이상 내가 나를 미워하지 않고 싫어하지 않을 때까지.</QuoteText>
-                        </QuoteBox>
-                        <BookIcon size="250" color="#4659A9"/>
-                    </EachSentence>
-                    <EachSentence onClick={()=> setModalIsOpen(true)}>
-                        <QuoteBox>
-                            <QuoteText>타인의 언어는 결코, 나의 정답이 될 수 없음을 알기에, 홀로 밤을 읽지 않기로 한다.</QuoteText>
-                        </QuoteBox>
-                        <BookIcon size="250" color="#6F61C6"/>
-                    </EachSentence>
-                    <EachSentence onClick={()=> setModalIsOpen(true)}>
-                        <QuoteBox>
-                            <QuoteText>슬픔을 병처럼 여기지 않겠다고 말하면서 나는 조금씩 의연해졌다. 슬픔에게도 비밀이 있을거라고. 그 비밀을 추궁하지 않기로 했다.</QuoteText>
-                        </QuoteBox>
-                        <BookIcon size="250" color="#2C2C60"/>
-                    </EachSentence>
+                    <EachSentence onClick={()=>setModalIsOpen(true)} quote="고독을 배설한 자리에서 내려앉는 환희. 이 달콤함을 위해 그는 예술을 표방한다." bookColor="#97A4E8" />
+                    <EachSentence onClick={()=>setModalIsOpen(true)} quote="전부 바다에 밀어버리자. 더 이상 내가 나를 미워하지 않고 싫어하지 않을 때까지." bookColor="#4659A9" />
+                    <EachSentence onClick={()=>setModalIsOpen(true)} quote="타인의 언어는 결코, 나의 정답이 될 수 없음을 알기에, 홀로 밤을 읽지 않기로 한다." bookColor="#6F61C6" />
+                    <EachSentence onClick={()=>setModalIsOpen(true)} quote="슬픔을 병처럼 여기지 않겠다고 말하면서 나는 조금씩 의연해졌다. 슬픔에게도 비밀이 있을거라고. 그 비밀을 추궁하지 않기로 했다." bookColor="#2C2C60" />
                 </SentencesContainer>
                 <BookModal 
                     isOpen={modalIsOpen} 
@@ -59,19 +30,7 @@ const TodaySentencePage = () => {
                             <BookImage/>
                             <BookInfo>
                                 <BookTitleBox>
-                                    <LikeIcon>
-                                        <svg width="0" height="0">
-                                            <linearGradient id="gradient" x1="100%" y1="0%" x2="100%" y2="100%">
-                                                <stop stopColor="#6F61C6" offset="0%" />
-                                                <stop stopColor="#97A4E8" offset="100%" />
-                                            </linearGradient>
-                                        </svg>
-                                        {isChecked ? (
-                                            <HeartFilledIcon size={45} onClick={onClick} />
-                                        ) : ( 
-                                            <HeartOutlinedIcon size={45} onClick={onClick} />
-                                        )}
-                                    </LikeIcon>
+                                    <LikeIcon/>
                                     <BookTitle>뼈가 자라는 여름</BookTitle>
                                 </BookTitleBox>
                                 <BookTextBox>
@@ -123,41 +82,6 @@ const ExplainationText = styled.p`
     font-family: "KOTRA_GOTHIC";
 `;
 
-const EachSentence = styled.div`
-    position: relative;
-    width: 500px;
-    margin: 70px 0;
-    margin-right: 20px;
-    margin-left: 90px;
-    padding: 0;
-`;
-
-const QuoteBox = styled.div`
-    position: relative;
-    align-items: center;
-    display: flex;
-    height: 147px;
-    width: 500px;
-    background-image: linear-gradient( to bottom, white, rgba(217, 217, 217, 0.5) );
-    border: 2px solid white;
-    clip-path: polygon(0% 0%, 100% 0, 88% 50%, 100% 100%, 0% 100%);
-`;
-
-const QuoteText = styled.div`
-    width: 300px;
-    margin: 0;
-    margin-left: 130px;
-    color: black;
-    font-size: 20px;
-    font-family: "KOTRA_GOTHIC";
-`;
-
-const BookIcon = styled(PiBookFill)`
-    position: absolute;
-    bottom: 5px;
-    right: 350px;
-`;
-
 const Content = styled.div`
     width: 1295px;
     height: 554px;
@@ -196,34 +120,12 @@ const BookTitleBox = styled.div`
     margin-bottom: 10px;
 `;
 
-const LikeIcon = styled.div`
-    margin-bottom: 5px;
-`;
-
-const HeartFilledIcon = styled(AiFillHeart)`
-    fill: url(#gradient);
-    cursor: pointer;
-    transition: transform 300ms ease;
-    &:hover {
-        transform: scale(1.1);
-    }
-`;
-
-const HeartOutlinedIcon = styled(AiOutlineHeart)`
-    color: #6F61C6;
-    cursor: pointer;
-    transition: transform 300ms ease;
-    &:hover {
-        transform: scale(1.1);
-    }
-`;
-
 const BookTitle = styled.p`
     font-size: 40px;
     font-family: "KOTRA_BOLD";
     margin: 0;
     margin-left: 10px;
-    
+    margin-top: 7px;
 `;
 
 const BookTextBox = styled.div`
@@ -234,7 +136,7 @@ const BookTextBox = styled.div`
     width: 665px;
     background-color: #2C2C60;
     clip-path: polygon(0% 0%, 100% 0, 90% 50%, 100% 100%, 0% 100%);
-    `;
+`;
 
 const BookSentence = styled.div`
     margin-left: 30px;
